pyproject.toml
[project]
name = "grovvy"
version = "0.1.0"
description = ""
authors = [
    {name = "Deyvid Augusto",email = "deyvidaugusto100@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "spotipy (>=2.25.1,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "openai (>=1.102.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "grovvy", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
